+++ PennyPack/ViewModels/CartViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  CartViewModel.swift
+//  PennyPack
+//
+//  Created by siye on 1/3/25.
+//
+
+import Foundation
+
+class CartViewModel: ObservableObject{
+    
+}
